+++ pkg/analyzer/lib/src/summary2/metadata_resolver.dart
@@ -24,8 +24,8 @@ class MetadataResolver extends ThrowingAstVisitor<void> {
   void visitAnnotation(covariant AnnotationImpl node) {
     node.elementAnnotation = ElementAnnotationImpl(_unitElement);
 
-    var astResolver = AstResolver(_linker, _unitElement, _scope);
-    astResolver.resolve(node, () => node);
+    var astResolver = AstResolver(_linker, _unitElement, _scope, node);
+    astResolver.resolveAnnotation(node);
   }
 
   @override