+++ pkg/dartdev/lib/src/templates/package_simple.dart
@@ -55,7 +55,7 @@ environment:
 #   path: ^1.8.0
 
 dev_dependencies:
-  pedantic: ^1.10.0
+  lints: ^1.0.0-0
   test: ^1.16.0
 ''';