+++ tests/language_2/mixin/illegal_cycles_test.dart
@@ -7,7 +7,7 @@
 class M {}
 class M0 extends Object with M0 { }
 //    ^^
-// [analyzer] COMPILE_TIME_ERROR.RECURSIVE_INTERFACE_INHERITANCE_WITH
+// [analyzer] COMPILE_TIME_ERROR.RECURSIVE_INTERFACE_INHERITANCE
 // [cfe] 'M0' is a supertype of itself.
 //    ^
 // [cfe] 'Object with M0' is a supertype of itself.
@@ -15,7 +15,7 @@ class M0 extends Object with M0 { }
 // [analyzer] COMPILE_TIME_ERROR.MIXIN_INHERITS_FROM_NOT_OBJECT
 class M1 = Object with M1;
 //    ^^
-// [analyzer] COMPILE_TIME_ERROR.RECURSIVE_INTERFACE_INHERITANCE_WITH
+// [analyzer] COMPILE_TIME_ERROR.RECURSIVE_INTERFACE_INHERITANCE
 // [cfe] 'M1' is a supertype of itself.
 
 class M2 = Object with M3;